requirements.txt
Flask==2.3.2
pymongo==4.4.0
python-dotenv==1.0.0
flask-cors==4.0.1
flask-socketio==5.3.6
   # optional for realtime; keep if you want sockets
python-engineio==4.5.0
requests==2.31.0
